+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Release 5.0.0
 
+- Add max auction sizes
 - Add ability to disable permissionless bids and restrict all trading to trusted fillers
+- Add ability to change name
 
 ## Release 4.0.0